requirements.txt
gspread==5.1.1
python-telegram-bot==13.10
toml==0.10.2
